MULE-16461: Remove `concatMap` in XML SDK chain execution
